renamed file cache-search to mem-cache
